refactor(ui): lift StateInspectorBody + DriftGauge to shared components
